refactor use 1 file for s3 caching
